requirements.txt
click==8.0.3
dnspython==2.2.0
Flask==2.0.2
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
pymongo==4.0.1
python-dotenv==0.19.2
Werkzeug==2.0.3
